chore: create kubelet endpoints controller
This commit moves the Kubelet endpoints reconciliation loop from the
Prometheus controller to a separate controller.
